// src/app/model/divida/divida.ts
export class Divida {
    
    idDivida: string;
    dsDespesa: string;
    dtVencimento: Date;
    nrDiaVencimento: number;
    nrMesVencimento: number;
    nrAnoVencimento: number;
    vlDespesa: number;
    idUsuario: number;
    fgDespesaPaga: boolean;
    fgPodePagar: boolean;
    nrDiaDiferenca: string;
    dsTipo: string;

}